pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "amolixs", email = "wilmet.amaury@protonmail.com"},
]
dependencies = [
    "pyside6>=6.4.2",
    "pytest>=7.2.1",
    "qt-custom-widgets>=0.1.6",
    "requests>=2.23.0",
]
requires-python = ">=3.10,<3.11"
license = {text = "MIT"}
